Move FileIO logic to ConfigFileLoader
