+++ Sources/SotoCore/Credential/ConfigFileLoader.swift
@@ -15,7 +15,11 @@
 import INIParser
 import Logging
 import NIO
-import struct Foundation.UUID
+#if os(Linux)
+import Glibc
+#else
+import Foundation.NSString
+#endif
 
 /// Load settings from AWS credentials and profile configuration files
 /// https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-files.html
@@ -66,16 +70,91 @@ struct ConfigFileLoader {
         case missingProfile(String)
         case missingAccessKeyId
         case missingSecretAccessKey
-        case missingSourceProfile
     }
 
-    /// Load profile configuraton from a file (passed in as byte-buffer), usually `~/.aws/config`
+    // MARK: - File IO
+
+    /// Load credentials from disk
+    /// - Parameters:
+    ///   - credentialsFilePath: file path for AWS credentials file
+    ///   - configFilePath: file path for AWS config file (optional)
+    ///   - profile: named profile to load (optional)
+    ///   - context: credential provider factory context
+    /// - Returns: Promise of a Credential Provider (StaticCredentials or STSAssumeRole)
+    static func loadSharedCredentials(
+        credentialsFilePath: String,
+        configFilePath: String?,
+        profile: String,
+        context: CredentialProviderFactory.Context
+    ) -> EventLoopFuture<(ProfileCredentials, ProfileConfig?)> {
+
+        let threadPool = NIOThreadPool(numberOfThreads: 1)
+        threadPool.start()
+        let fileIO = NonBlockingFileIO(threadPool: threadPool)
+
+        return loadFile(path: credentialsFilePath, on: context.eventLoop, using: fileIO)
+            .always { _ in
+                // shutdown the threadpool async
+                threadPool.shutdownGracefully { _ in }
+            }
+            .flatMap { credentialsByteBuffer -> EventLoopFuture<(ByteBuffer, ByteBuffer?)> in
+                if let path = configFilePath {
+                    return loadFile(path: path, on: context.eventLoop, using: fileIO).map { (credentialsByteBuffer, $0) }
+                }
+                return context.eventLoop.makeSucceededFuture((credentialsByteBuffer, nil))
+            }
+            .flatMapThrowing { credentialsByteBuffer, configByteBuffer in
+                return try parseSharedCredentials(from: credentialsByteBuffer, configByteBuffer: configByteBuffer, for: profile)
+            }
+    }
+
+    /// Load a file from disk without blocking the current thread
+    /// - Parameters:
+    ///   - path: path for the file to load
+    ///   - eventLoop: event loop to run everything on
+    ///   - fileIO: non-blocking file IO
+    /// - Returns: Event loop future with file contents in a byte-buffer
+    static func loadFile(path: String, on eventLoop: EventLoop, using fileIO: NonBlockingFileIO) -> EventLoopFuture<ByteBuffer> {
+        let path = expandTildeInFilePath(path)
+
+        return fileIO.openFile(path: path, eventLoop: eventLoop)
+            .flatMap { handle, region in
+                fileIO.read(fileRegion: region, allocator: ByteBufferAllocator(), eventLoop: eventLoop).map { ($0, handle) }
+            }
+            .flatMapThrowing { byteBuffer, handle in
+                try handle.close()
+                return byteBuffer
+            }
+    }
+
+    // MARK: - Byte Buffer parsing (INIParser)
+
+    /// Parse credentials from files (passed in as byte-buffers)
+    /// 
+    /// Credentials file settings have precedence over profile configuration settings
+    /// https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-quickstart.html#cli-configure-quickstart-precedence
+    ///
+    /// - Parameters:
+    ///   - credentialsBuffer: contents of AWS shared credentials file (usually `~/.aws/credentials`
+    ///   - configByteBuffer: contents of AWS profile configuration file (usually `~/.aws/config`
+    ///   - profile: named profile to load (optional)
+    /// - Returns: Promise of a Credential Provider (StaticCredentials or STSAssumeRole)
+    static func parseSharedCredentials(from credentialsByteBuffer: ByteBuffer, configByteBuffer: ByteBuffer?, for profile: String) throws -> (ProfileCredentials, ProfileConfig?) {
+        var config: ProfileConfig?
+        if let byteBuffer = configByteBuffer {
+            config = try parseProfileConfig(from: byteBuffer, for: profile)
+        }
+        let credentials = try parseCredentials(from: credentialsByteBuffer, for: profile, sourceProfile: config?.sourceProfile)
+        return (credentials, config)
+    }
+
+    /// Parse profile configuraton from a file (passed in as byte-buffer), usually `~/.aws/config`
     ///
     /// - Parameters:
     ///   - byteBuffer: contents of the file to parse
     ///   - profile: AWS named profile to load (usually `default`)
     /// - Returns: Combined profile settings
-    static func loadProfileConfig(from byteBuffer: ByteBuffer, for profile: String) throws -> ProfileConfig {
+    static func parseProfileConfig(from byteBuffer: ByteBuffer, for profile: String) throws -> ProfileConfig {
         guard let content = byteBuffer.getString(at: 0, length: byteBuffer.readableBytes) else {
             throw ConfigFileError.invalidCredentialFileSyntax
         }
@@ -106,16 +185,16 @@ struct ConfigFileLoader {
         )
     }
 
-    /// Load profile credentials from a file (passed in as byte-buffer), usually `~/.aws/credentials`
+    /// Parse profile credentials from a file (passed in as byte-buffer), usually `~/.aws/credentials`
     ///
-    /// If `source_profile` is passed in, or found in the settings, credentials will be loaded from the source profile for use with STS Assume Role.
+    /// If `source_profile` is passed in, or found in the settings, credentials will be loaded from the source profile for using with STS Assume Role.
     ///
     /// - Parameters:
     ///   - byteBuffer: contents of the file to parse
     ///   - profile: AWS named profile to load (usually `default`)
     ///   - sourceProfile: specifies a named profile with long-term credentials that the AWS CLI can use to assume a role that you specified with the `role_arn` parameter.
     /// - Returns: Combined profile credentials
-    static func loadCredentials(from byteBuffer: ByteBuffer, for profile: String, sourceProfile: String?) throws -> ProfileCredentials {
+    static func parseCredentials(from byteBuffer: ByteBuffer, for profile: String, sourceProfile: String?) throws -> ProfileCredentials {
         guard let content = byteBuffer.getString(at: 0, length: byteBuffer.readableBytes) else {
             throw ConfigFileError.invalidCredentialFileSyntax
         }
@@ -161,6 +240,42 @@ struct ConfigFileLoader {
             sourceProfile: sourceProfile ?? settings["source_profile"],
             credentialSource: settings["credential_source"].flatMap(CredentialSource.init(rawValue:))
         )
+    }
+
+    // MARK: - Path Expansion
+
+    static func expandTildeInFilePath(_ filePath: String) -> String {
+        #if os(Linux)
+        // We don't want to add more dependencies on Foundation than needed.
+        // For this reason we get the expanded filePath on Linux from libc.
+        // Since `wordexp` and `wordfree` are not available on iOS we stay
+        // with NSString on Darwin.
+        return filePath.withCString { (ptr) -> String in
+            var wexp = wordexp_t()
+            guard wordexp(ptr, &wexp, 0) == 0, let we_wordv = wexp.we_wordv else {
+                return filePath
+            }
+            defer {
+                wordfree(&wexp)
+            }
+
+            guard let resolved = we_wordv[0], let pth = String(cString: resolved, encoding: .utf8) else {
+                return filePath
+            }
+
+            return pth
+        }
+        #elseif os(macOS)
+        // can not use wordexp on macOS because for sandboxed application wexp.we_wordv == nil
+        guard let home = getpwuid(getuid())?.pointee.pw_dir,
+            let homePath = String(cString: home, encoding: .utf8)
+        else {
+            return filePath
+        }
+        return filePath.starts(with: "~") ? homePath + filePath.dropFirst() : filePath
+        #else
+        return NSString(string: filePath).expandingTildeInPath
+        #endif
     }
 
 }
